requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
oauthlib==3.2.0
pkg-resources==0.0.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
tweepy==4.6.0
urllib3==1.26.11
